* Properly detect issues with 'd.directory.set' being called after the data files have been moved.
* Added counters for unchoked peers.

* CLang fixes.


git-svn-id: svn://rakshasa.no/libtorrent/trunk/libtorrent@1272 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/protocol/extensions.cc
@@ -220,7 +220,7 @@ ProtocolExtension::read_start(int type, uint32_t length, bool skip) {
   if (is_default() || (type >= FIRST_INVALID) || length > (1 << 15))
     throw communication_error("Received invalid extension message.");
 
-  if (m_read != NULL || length < 0)
+  if (m_read != NULL || (int32_t)length < 0)
     throw internal_error("ProtocolExtension::read_start called in inconsistent state.");
 
   m_readLeft = length;
